refactoring for asset requests
