+++ src/pages/interactive-exercise/[slug]/index.jsx
@@ -44,22 +44,14 @@ import GridContainer from '../../../common/components/GridContainer';
 import redirectsFromApi from '../../../../public/redirects-from-api.json';
 // import MktSideRecommendedCourses from '../../../common/components/MktSideRecommendedCourses';
 import useStyle from '../../../common/hooks/useStyle';
-import { parseQuerys } from '../../../utils/url';
 import { cleanObject } from '../../../utils';
-import { ORIGIN_HOST, WHITE_LABEL_ACADEMY } from '../../../utils/variables';
+import { ORIGIN_HOST } from '../../../utils/variables';
+import { getAsset } from '../../../utils/requests';
 
 export const getStaticPaths = async ({ locales }) => {
-  const querys = parseQuerys({
-    asset_type: 'EXERCISE',
-    visibility: 'PUBLIC',
-    status: 'PUBLISHED',
-    academy: WHITE_LABEL_ACADEMY,
-    limit: 2000,
-  });
-  const resp = await fetch(`${process.env.BREATHECODE_HOST}/v1/registry/asset${querys}`);
-  const data = await resp.json();
+  const data = await getAsset('EXERCISE', {});
 
-  const paths = data.results.flatMap((res) => locales.map((locale) => ({
+  const paths = data.flatMap((res) => locales.map((locale) => ({
     params: {
       slug: res.slug,
     },
